chore(core): endurecer tipos y apagar warnings heredados
Normaliza conversiones estrechas en core/clientes (puertos, colas, contadores, hashsets) para eliminar C4242 y documenta el cierre en el plan de modernización. Congela los lexers de flex por defecto con la opción WMULE_USE_FLEX, limpia register/strdup del Scanner pre-generado y evita los C4005 de PACKAGE encapsulando los defines en config.h. Se actualiza la documentación y se mantiene la build/test suite en verde tras los cambios.

+++ src/KnownFileList.h
@@ -51,7 +51,7 @@ public:
 	void	ReleaseIndex();
 
 	uint16 requested;
-	uint32 transferred;
+	uint64 transferred;
 	uint16 accepted;
 
 private:
